Add recipe for `RotatedGrid` (#704)
* Initial implementation

* Add recipe for 'RotatedGrid'

* Use the 'boundingbox' function
